+++ src/main/java/edu/dosw/project/SFC_TechUp_Futbol/persistence/mapper/OrganizadorMapper.java
@@ -18,7 +18,7 @@ public class OrganizadorMapper {
         entity.setPassword(organizador.getPassword());
         entity.setUserType(organizador.getUserType());
         if (organizador.getCurrentTournament() != null) {
-            entity.setTorneoId((long) organizador.getCurrentTournament().getId());
+            entity.setTorneoId(organizador.getCurrentTournament().getId());
         }
         return entity;
     }
